refactor: Update AppButton to accept onTap and title parameters
The AppButton widget in app_button.dart has been updated to accept two new parameters: onTap and title. This allows for more flexibility in using the button widget throughout the application. The onTap parameter is a callback function that will be executed when the button is tapped, and the title parameter is a string that represents the text displayed on the button.

This commit refactors the AppButton widget to incorporate these changes, making it more versatile and customizable.